+++ src/Spinner.tsx
@@ -1,0 +1,13 @@
+import React from "react";
+
+class Spinner extends React.Component {
+  render() {
+    return (
+      <div className="spinner-container">
+        <div className="spinner"></div>
+      </div>
+    );
+  }
+}
+
+export default Spinner;
